Add "Shodan" Pattern Type to Indicators (#2446)
