+++ opencti-platform/opencti-front/src/utils/Localization.js
@@ -1520,6 +1520,7 @@ const i18n = {
       'tanium-signal': 'Tanium Signal',
       spl: 'Splunk SPL',
       eql: 'Elastic EQL',
+      shodan: 'Shodan',
       'Apply on': 'Aplicar sobre',
       'Retention policies': 'Política de retención',
       'Max retention': 'Retención máxima',
@@ -3062,6 +3063,7 @@ const i18n = {
       'tanium-signal': 'Tanium Signal',
       spl: 'Splunk SPL',
       eql: 'Elastic EQL',
+      shodan: 'Shodan',
       'Apply on': "S'applique à",
       'Retention policies': 'Politiques de rétention',
       'Max retention': 'Rétention maximum',
@@ -4579,6 +4581,7 @@ const i18n = {
       'tanium-signal': 'Taniumシグナル',
       spl: 'Splunk SPL',
       eql: 'Elastic EQL',
+      shodan: 'Shodan',
       'Apply on': '適用',
       'Retention policies': '保持ポリシー',
       'Max retention': '最大保持数',
@@ -6002,6 +6005,7 @@ const i18n = {
       'tanium-signal': 'Tanium信号',
       spl: 'Splunk SPL',
       eql: 'Elastic EQL',
+      shodan: 'Shodan',
       'Apply on': '应用于',
       'Retention policies': '保留策略',
       'Max retention': '最长保留期',
@@ -6314,6 +6318,7 @@ const i18n = {
       'tanium-signal': 'Tanium Signal',
       spl: 'Splunk SPL',
       eql: 'Elastic EQL',
+      shodan: 'Shodan',
       'Received time': 'Triggered',
       'Processed time': 'Data sent',
       true: 'Yes',
